pyproject.toml
[tool.poetry]
name = "healthy habits tracker"
version = "0.1.0"
description = ""
authors = ["Yana Sozonova <ynakochneva@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.2"
black = "^25.1.0"
isort = "^6.0.1"
django = "^5.1.7"
dotenv = "^0.9.9"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.10"
celery = "^5.4.0"
redis = "^5.2.1"
eventlet = "^0.39.1"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
coverage = "^7.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '\.git'


[tool.isort]
line_length = 119
